Significant refactoring of the results pages
